docs: split docs for REACT-VFX and VFX-JS
