+++ packages/docs-vfx-js/index.html
@@ -1,0 +1,148 @@
+<!doctype html>
+<html lang="en">
+    <head>
+        <meta charset="utf-8" />
+        <meta name="viewport" content="width=device-width, initial-scale=1" />
+
+        <title>REACT-VFX - WebGL effects for React! #REACTVFX</title>
+
+        <meta name="theme-color" content="#000000" />
+        <meta
+            name="description"
+            content="REACT-VFX is a React component library which allows you to add WebGL powered effects to you React aaplication."
+        />
+        <link rel="apple-touch-icon" href="logo-with-bg.png" />
+
+        <meta property="og:url" content="https://amagi.dev/react-vfx/" />
+        <meta property="og:type" content="website" />
+        <meta
+            property="og:title"
+            content="REACT-VFX - WebGL effects for React! #REACTVFX"
+        />
+        <meta property="og:image" content="logo-with-bg.png" />
+        <meta
+            property="og:description"
+            content="REACT-VFX is a React component library which allows you to add WebGL powered effects to you React aaplication."
+        />
+        <meta
+            property="og:site_name"
+            content="REACT-VFX - WebGL effects for React! #REACTVFX"
+        />
+        <meta property="og:locale" content="en_US" />
+
+        <meta name="twitter:card" content="summary_large_image" />
+        <meta name="twitter:site" content="@amagitakayosi" />
+        <meta name="twitter:creator" content="@amagitakayosi" />
+        <meta name="twitter:url" content="https://amagi.dev/react-vfx/" />
+        <meta
+            name="twitter:title"
+            content="REACT-VFX - WebGL effects for React! #REACTVFX"
+        />
+        <meta
+            name="twitter:description"
+            content="REACT-VFX is a React component library which allows you to add WebGL powered effects to you React aaplication."
+        />
+        <meta
+            name="twitter:image"
+            content="https://amagi.dev/react-vfx/logo-with-bg.png"
+        />
+
+        <link rel="icon" href="favicon/favicon.ico" />
+        <link
+            rel="icon"
+            type="image/x-icon"
+            sizes="16x16 32x32"
+            href="favicon/favicon-36.png"
+        />
+        <link
+            rel="apple-touch-icon"
+            sizes="180x180"
+            href="favicion/favicon-180.png"
+        />
+        <link rel="icon" sizes="192x192" href="favicon/favicon-192.png" />
+
+        <link rel="stylesheet" href="./index.css" />
+    </head>
+
+    <body>
+        <noscript>You need to enable JavaScript to run this app.</noscript>
+
+        <header>
+            <div class="HeaderTop">
+                <a
+                    href="https://github.com/fand/react-vfx/tree/main/packages/vfx-js"
+                    target="_blank"
+                    >GITHUB</a
+                >
+            </div>
+            <!-- BG for legibility -->
+            <div class="HeaderBottom">GITHUB</div>
+        </header>
+
+        <img
+            id="BG"
+            src="./vfx-js-blob.webp"
+            alt="VFX-JS"
+            data-shader-id="blob"
+        />
+
+        <section class="TopSection">
+            <img
+                class="Logo"
+                src="../vfx-js-logo-no-padding.svg"
+                alt="VFX-JS"
+                data-shader="rgbShift"
+                data-overflow="100"
+            />
+
+            <p data-shader="rgbShift" data-overflow="100">
+                VFX-JS is a JavaScript library to add WebGL-powered effects to
+                your website. You can easily attach it to normal
+                <code> &lt;img /&gt;</code>, <code>&lt;video&gt;</code> elements
+                etc.
+            </p>
+        </section>
+
+        <section class="ExampleSection">
+            <h2>Examples</h2>
+            <img src="./vfxjs-logo.png" data-shader="rainbow" />
+            <img src="./david.png" data-shader="rgbShift" />
+            <img src="./cat.gif" data-shader="rainbow" />
+            <img src="./chill.gif" data-shader="sinewave" />
+            <img src="./doge.gif" data-shader="pixelate" />
+            <img src="./octocat.gif" data-shader="glitch" />
+
+            <video src="../mind_blown.mp4" autoplay loop muted></video>
+        </section>
+
+        <section class="UsageSection">
+            <h2>Usage</h2>
+
+            <p>Install via npm:</p>
+            <pre><code>
+npm i vfx-js
+            </code></pre>
+
+            <p>Then create <code>VFX</code> object in your script:</p>
+            <pre><code>
+import { VFX } from 'vfx-js';
+const vfx = new VFX();
+vfx.addImage(imgElement, { shader: "rgbShift" });
+            </code></pre>
+
+            <p>This will be rendered as below.</p>
+            <img src="./react-logo.png" data-shader="rgbShift" />
+        </section>
+
+        <section class="AuthorSection">
+            <img src="./amagi.png" />
+            <h2>Made by Amagi</h2>
+        </section>
+
+        <div id="MaskBottom"></div>
+        <div id="Frame"></div>
+        <div id="MaskTop"></div>
+
+        <script type="module" src="/src/main.ts"></script>
+    </body>
+</html>
