+++ packages/docs-vfx-js/public/index.html
@@ -1,0 +1,45 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <meta charset="utf-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1" />
+
+    <title>REACT-VFX - WebGL effects for React! #REACTVFX</title>
+
+    <meta name="theme-color" content="#000000" />
+    <meta name="description"
+        content="REACT-VFX is a React component library which allows you to add WebGL powered effects to you React aaplication." />
+    <link rel="apple-touch-icon" href="%PUBLIC_URL%/logo-with-bg.png" />
+
+    <meta property="og:url" content="https://amagi.dev/react-vfx/" />
+    <meta property="og:type" content="website" />
+    <meta property="og:title" content="REACT-VFX - WebGL effects for React! #REACTVFX" />
+    <meta property="og:image" content="%PUBLIC_URL%/logo-with-bg.png" />
+    <meta property="og:description"
+        content="REACT-VFX is a React component library which allows you to add WebGL powered effects to you React aaplication." />
+    <meta property="og:site_name" content="REACT-VFX - WebGL effects for React! #REACTVFX" />
+    <meta property="og:locale" content="en_US" />
+
+    <meta name="twitter:card" content="summary_large_image" />
+    <meta name="twitter:site" content="@amagitakayosi" />
+    <meta name="twitter:creator" content="@amagitakayosi" />
+    <meta name="twitter:url" content="https://amagi.dev/react-vfx/" />
+    <meta name="twitter:title" content="REACT-VFX - WebGL effects for React! #REACTVFX" />
+    <meta name="twitter:description"
+        content="REACT-VFX is a React component library which allows you to add WebGL powered effects to you React aaplication." />
+    <meta name="twitter:image" content="https://amagi.dev/react-vfx/logo-with-bg.png" />
+
+    <link rel="icon" href="%PUBLIC_URL%/favicon/favicon.ico" />
+    <link rel="icon" type="image/x-icon" sizes="16x16 32x32" href="%PUBLIC_URL%/favicon/favicon-36.png" />
+    <link rel="apple-touch-icon" sizes="180x180" href="%PUBLIC_URL%/favicion/favicon-180.png" />
+    <link rel="icon" sizes="192x192" href="%PUBLIC_URL%/favicon/favicon-192.png" />
+
+</head>
+
+<body>
+    <noscript>You need to enable JavaScript to run this app.</noscript>
+    <div id="root"></div>
+</body>
+
+</html>
